+++ tests/tests/__init__.py
@@ -1,0 +1,2 @@
+from .cim import *
+from .views import *
